readme

+++ Qt/glwidget.cc
@@ -118,13 +118,12 @@ void GLWidget::timerEvent(QTimerEvent* event)
   Q_UNUSED(event);
 
 
-  double dt = chronometre.restart() / (double) constantes::time_division;
+  double dt = chronometre.restart() / constantes::time_division;
 
 
   for(int i = 0; i < constantes::turn_before_update; i++){
        acc.evolue(dt);
   }
-
   update();
 }
